+++ src/Lurog.NET/Lurog.NET/Logger.cs
@@ -2,9 +2,17 @@
 {
     public class Logger : ILogger
     {
-        public void Log(object sender, ErrorRaisedEventArgs e)
+        /*
+         * This is a basic console log method which can be overridden by deriving classes.
+         * Deriving classes may include, but aren't limited to: File Loggers, DB Loggers, message box loggers.
+         */
+
+        public virtual void Log(object sender, ErrorRaisedEventArgs e)
         {
-            throw new NotImplementedException();
+            Console.WriteLine($"Error Code: {e.ErrorCode}");
+            Console.WriteLine($"Severity: {e.Severity}");
+            Console.WriteLine($"Date & Time of occurence: {e.TimeOccurence}");
+            Console.WriteLine($"Message: {e.Message}\n");
         }
     }
 }